ingest_module/pyproject.toml
[project]
name = "ingest-module"
version = "0.1.0"
description = "Python app to ingest data from the IMDb datasource and store it in a Postgresql database"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.0.0",
    "pandas>=2.2.0",
    "requests>=2.32.0",
    "sqlalchemy>=2.0.0",
    "python-dateutil>=2.8.0",
    "psycopg2-binary>=2.9.0",
    "pytest>=9.0.2",
    "alive-progress>=3.3.0",
]
